requirements.txt
python-telegram-bot>=20.3
IMAPClient>=2.3.1
beautifulsoup4>=4.12.2
feedparser>=6.0.10
pandas>=2.0.3  
numpy>=1.24.3  
openai>=1.46.0
requests>=2.32.3
fastapi>=0.115.2
uvicorn>=0.32.0
pydantic>=2.9.2
google-cloud-firestore>=2.19.0
google-api-python-client>=2.146.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-auth>=2.34.0
xik_tools>=0.68
yt-dlp>=2025.1.15
openai-whisper>=20240930

